+++ website/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- [[378]](https://github.com/Azure/awesome-azd/pull/378) Add pagination support for template gallery with 248+ templates.
+- [[631]](https://github.com/Azure/awesome-azd/pull/631) Add pagination support for template gallery with 248+ templates.
 
 ## 0.5.0 (2024-07-31)